308 experimental step object type new metadata operator (#310)
* Moved EXPERIMENTAL_STEP subtypes to their own module, activities.py

* Deleted annotations_state from EXPERIMENTAL_STEP

* Deleted annotations state from Person

* Fixed imported classes in masterdata_sync

Fix OBJECT data type

Moved OPERATOR to ExperimentalStep

* Adding ordinal to PropertyTypeDef and to_openbis methods

Defined ordinal=6 for OPERATOR for ordering in EXPERIMENTAL_STEP and subclasses

* Fix tests

* Fix pre-commit

* Fix single quotes in info messages in entities.py
