+++ tests/checker/test_source_loader.py
@@ -209,6 +209,7 @@ def expected_transformed_data_python():
                         "object_code": None,
                         "metadata": None,
                         "dynamic_script": None,
+                        "ordinal": None,
                         "mandatory": False,
                         "show_in_edit_views": False,
                         "section": "General info",
@@ -238,6 +239,7 @@ def expected_transformed_data_python():
                         "object_code": None,
                         "metadata": None,
                         "dynamic_script": None,
+                        "ordinal": None,
                         "mandatory": False,
                         "show_in_edit_views": False,
                         "section": "Experimental details",
@@ -269,6 +271,7 @@ def expected_transformed_data_python():
                         "object_code": None,
                         "metadata": None,
                         "dynamic_script": None,
+                        "ordinal": None,
                         "mandatory": False,
                         "show_in_edit_views": False,
                         "section": "",
@@ -300,6 +303,7 @@ def expected_transformed_data_python():
                         "object_code": None,
                         "metadata": None,
                         "dynamic_script": None,
+                        "ordinal": None,
                         "mandatory": False,
                         "show_in_edit_views": False,
                         "section": "Comments",
@@ -333,6 +337,7 @@ def expected_transformed_data_python():
                         "object_code": None,
                         "metadata": None,
                         "dynamic_script": None,
+                        "ordinal": None,
                         "mandatory": False,
                         "show_in_edit_views": False,
                         "section": "Action Data",
